// frontend/src/api/generated/models/attendanceResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface AttendanceResponse {
  id: string;
  operator_id: string;
  operator_name: string;
  date: string;
  is_present: boolean;
  notes: string | null;
}